[refact] nonk8s-filter 분리
